requirements.txt
langchain
langchain-openai
langchain-community
langgraph
openai
google-genai
python-dotenv
rich